Use the provided groupID as the grpcomm signature
If the user provides a groupID (e.g., in a PMIx_Group op),
then use that as the signature of any involved grpcomm
operations instead of the proc array. This allows the same
collection of procs to participate in multiple PMIx_Group
operations at the same time since PMIx requires the user
provide a unique grpID for each operation.

Signed-off-by: Ralph Castain <rhc@pmix.org>