various fixes for singleton support
some recent changes broke singleton support - twice in a month.

First, remove problematic PMIX_RELEASE of jdata when its not ready to be removed.

For some reason this showed up in singleton mode with debug enabled.
Various asserts would fail when this PMIX_RELEASE was invoked.
This was due to the fact that the jdata had been put on a list of jdata's
so the opal_list destructor was having a fit trying to release a jdata
which was still in a list.

It turns out this jdata is being released in the code starting at
line 95 of prte_finalize.c.   I assume with debug not enabled that the
jdata is released twice, rather than failing in the assert in prted_comm.c

Some work to add in session id's for tracking allocations also broke
singleton support.

This patch restores the singletone functionality.

Related to issue https://github.com/open-mpi/ompi/issues/12307

Signed-off-by: Howard Pritchard <howardp@lanl.gov>